+++ src/slices/index.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { authReducer } from './authSlice';
-import { productCardsReducer } from './ProductCardsSlice';
+import { productCardsReducer } from './productCardsSlice';
 import { categoriesReducer } from './categoriesSlice';
 import { cartReducer } from './cartSlice';
