+++ src/utils/routesConfig.tsx
@@ -6,7 +6,7 @@ import { NotFoundPage } from '../pages/NotFoundPage';
 import { NavRoutes } from './routes';
 import { AboutPage } from '../pages/AboutPage';
 import { ContactsPage } from '../pages/ContactsPage';
-import { Product } from '../pages/ProductPage';
+import { Product } from '../pages/ProductPage/ProductPage';
 import { CatalogPage } from '../pages/catalogPage/CatalogPage';
 import { UserProfilePage } from '../pages/UserProfilePage';
 import { AnonymOutlet } from '../components/AnonymOutlet';
